pyproject.toml
[project]
name = "arcagi3"
version = "0.1.0"
description = "Clean benchmarking framework for ARC-AGI-3 games"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}

dependencies = [
    "pydantic>=2.0",
    "requests>=2.31.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "anthropic>=0.21.0",
    "openai>=1.12.0",
    "google-generativeai>=0.4.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
arcagi3 = ["*.yml"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

